update event page styles and API integration
