+++ shared/src/hooks/polling.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { GettablePaths, TypedAxiosInstance } from "../api/type-wrapper";
 import { APIRoutes } from "../api/types";
 
@@ -13,29 +13,33 @@ const usePolling = <T extends GettablePaths>(
   const [error, setError] = useState<string | null>(null);
   const isLoading = data === null && error === null;
 
-  useEffect(() => {
-    const fetchData = () => {
-      api
-        .get(endpoint)
-        .then((response) => {
-          setData(response.data);
-        })
-        .catch((err: { error: string }) => {
-          setError(err.error);
-        });
-    };
+  const fetchData = useCallback(() => {
+    api
+      .get(endpoint)
+      .then((response) => {
+        setData(response.data);
+      })
+      .catch((err: { error: string }) => {
+        setError(err.error);
+      });
+  }, [api, endpoint]);
 
+  useEffect(() => {
     fetchData();
     const id = setInterval(fetchData, interval);
 
     return () => clearInterval(id);
-  }, [api, endpoint, interval]);
+  }, [fetchData, interval]);
 
   const mutate = (updater: (prev: typeof data) => typeof data) => {
     setData((prev) => updater(prev));
   };
 
-  return { data, error, isLoading, mutate };
+  const update = () => {
+    fetchData();
+  };
+
+  return { data, error, isLoading, mutate, update };
 };
 
 export default usePolling;
